rename components
