+++ src/components/MainPage/MainPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 // import PropTypes from 'prop-types';
-import * as S from './styled.component';
+import * as S from './styled';
 
 const MainPage = () => {
   return <S.MainPageWrapper></S.MainPageWrapper>;
