+++ src/components/MainPage/styled.jsx
