+++ src/components/app.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Footer from './footer/footer.component';
-import MainPage from './mainPage/mainPage.component';
-import Textbook from './textbook/textbook.component';
+import Footer from './Footer';
 import Header from './Header';
+import MainPage from './MainPage';
+import Textbook from './Textbook';
 
 const App = () => {
   return (
