+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/app.component';
+import App from './components/app';
 
 import './assets/scss/style.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
